discard useless configurations; correct words
I won't offer a user or myself a fake description on a simple behavior of programs
https://en.wikipedia.org/wiki/Strategy
In the sense of the "art of the general", which included several subsets of skills including military tactics, siegecraft, logistics etc., the term came into use in the 6th century C.E. in Eastern Roman terminology, and was translated into Western vernacular languages only in the 18th century. From then until the 20th century, the word "strategy" came to denote "a comprehensive way to try to pursue political ends, including the threat or actual use of force, in a dialectic of wills" in a military conflict, in which both adversaries interact.

BTW, formulation and implementation are anywhere in this program

https://github.com/DNSCrypt/dnscrypt-proxy/issues/1555
https://github.com/DNSCrypt/dnscrypt-proxy/issues/1558
https://github.com/DNSCrypt/dnscrypt-proxy/issues/1559
https://github.com/DNSCrypt/dnscrypt-proxy/issues/1561
https://github.com/DNSCrypt/dnscrypt-proxy/issues/1562